Change SearchAsync<T> signature
The old signature was a bit akward, it took as first paramater a string for the query and a second parameter an optional SearchQuery object where we store later the first parameter into the Q property of the SearchQuery.

The main issue was that the SearchQuery can be created with the default constructor which is wrong. The SearchQuery object must be created by passing a search query to the constructor to be in a valid state.